+++ programs/gmsol-store/src/states/order.rs
@@ -409,13 +409,17 @@ impl Closable for Order {
 
 impl Order {
     /// Get rent for position cut.
-    pub fn position_cut_rent(is_pure: bool) -> Result<u64> {
+    pub(crate) fn position_cut_rent(is_pure: bool, include_execution_fee: bool) -> Result<u64> {
         use anchor_spl::token::TokenAccount;
 
         let rent = Rent::get()?;
         let amount = rent.minimum_balance(Self::INIT_SPACE + 8)
             + rent.minimum_balance(TokenAccount::LEN) * if is_pure { 1 } else { 2 }
-            + Self::MIN_EXECUTION_LAMPORTS;
+            + if include_execution_fee {
+                Self::MIN_EXECUTION_LAMPORTS
+            } else {
+                0
+            };
 
         Ok(amount)
     }